chore: sortir les fonction qui anonymise les followup et simulation
