+++ tools/cleaner.ts
@@ -4,99 +4,10 @@ import config from "../backend/config/index.js"
 import mongooseConfig from "../backend/config/mongoose.js"
 import Simulation from "../backend/models/simulation"
 import Followup from "../backend/models/followup"
-
-function getAnonymizedAnswer(answer, simulation) {
-  switch (answer.entityName) {
-    case "famille": {
-      switch (answer.fieldName) {
-        case "en_couple": {
-          return answer
-        }
-        default: {
-          return null
-        }
-      }
-    }
-    case "menage": {
-      switch (answer.fieldName) {
-        case "depcom":
-        case "loyer":
-        case "statut_occupation_logement":
-        case "_logementType":
-        case "_primoAccedant":
-        case "_locationType": {
-          return answer
-        }
-        default: {
-          return null
-        }
-      }
-    }
-    case "individu": {
-      if (answer.id == "demandeur" || answer.id == "conjoint") {
-        switch (answer.fieldName) {
-          case "_contrat_alternant":
-          case "_hasRessources":
-          case "_interetAidesSanitaireSocial":
-          case "_interetBafa":
-          case "_interetEtudesEtranger":
-          case "_interetPermisDeConduire":
-          case "_interetsAidesVelo":
-          case "activite":
-          case "age":
-          case "alternant":
-          case "boursier":
-          case "depcom":
-          case "enfant_a_charge":
-          case "nationalite":
-          case "ressources": {
-            return answer
-          }
-          case "date_naissance": {
-            const dt = new Date(simulation.dateDeValeur)
-            const dob = new Date(answer.value)
-            return {
-              id: answer.id,
-              entityName: answer.entityName,
-              fieldName: "age",
-              value: Math.round(
-                (dt.getTime() - dob.getTime()) / 365.25 / 24 / 60 / 60 / 1000
-              ),
-            }
-          }
-          default: {
-            return null
-          }
-        }
-      } else {
-        return null
-      }
-    }
-    default: {
-      return null
-    }
-  }
-}
-
-function anonymizeSimulation(simulation) {
-  const answers = simulation.answers.all
-  const answersAnonymized = answers.map((answer) => getAnonymizedAnswer(answer, simulation)).filter((answer) => answer)
-
-  simulation.answers = {
-    all: answersAnonymized,
-    current: [],
-  }
-  simulation.status = "anonymized"
-
-  return simulation
-}
-
-function anonymizeFollowup(followup) {
-  followup.email = undefined
-  followup.error = undefined
-
-  return followup
-}
+import {
+  anonymizeSimulation,
+  anonymizeFollowup,
+} from "../lib/cleaner-functions"
 
 async function main() {
   const aMonthAgo = Date.now() - 31 * 24 * 60 * 60 * 1000
